edited empty list case
